🍑 fix static file not read and add tag_list, category_list function

+++ src/main/java/com/tale/service/impl/ContentsServiceImpl.java
@@ -58,7 +58,7 @@ public class ContentsServiceImpl implements ContentsService {
         if (StringKit.isBlank(contents.getTitle()))
             throw new TipException("文章标题不能为空");
         if (contents.getTitle().length() > TaleConst.MAX_TITLE_COUNT) {
-            throw new TipException("文章标题最多可以输入"+ TaleConst.MAX_TITLE_COUNT +"个字符");
+            throw new TipException("文章标题最多可以输入" + TaleConst.MAX_TITLE_COUNT + "个字符");
         }
 
         if (StringKit.isBlank(contents.getContent()))
@@ -66,7 +66,7 @@ public class ContentsServiceImpl implements ContentsService {
         // 最多可以输入5w个字
         int len = contents.getContent().length();
         if (len > TaleConst.MAX_TEXT_COUNT)
-            throw new TipException("文章内容最多可以输入"+ TaleConst.MAX_TEXT_COUNT +"个字符");
+            throw new TipException("文章内容最多可以输入" + TaleConst.MAX_TEXT_COUNT + "个字符");
         if (null == contents.getAuthor_id())
             throw new TipException("请登录后发布文章");
 
@@ -106,18 +106,17 @@ public class ContentsServiceImpl implements ContentsService {
             throw new TipException("文章标题不能为空");
         }
         if (contents.getTitle().length() > TaleConst.MAX_TITLE_COUNT) {
-            throw new TipException("文章标题最多可以输入"+ TaleConst.MAX_TITLE_COUNT +"个字符");
+            throw new TipException("文章标题最多可以输入" + TaleConst.MAX_TITLE_COUNT + "个字符");
         }
         if (StringKit.isBlank(contents.getContent())) {
             throw new TipException("文章内容不能为空");
         }
         if (contents.getContent().length() > TaleConst.MAX_TEXT_COUNT)
-            throw new TipException("文章内容最多可以输入"+ TaleConst.MAX_TEXT_COUNT +"个字符");
+            throw new TipException("文章内容最多可以输入" + TaleConst.MAX_TEXT_COUNT + "个字符");
         if (null == contents.getAuthor_id()) {
             throw new TipException("请登录后发布文章");
         }
-        int time = (int) DateKit.nowUnix();
-        contents.setModified(time);
+        contents.setModified(DateKit.nowUnix());
 
         Integer cid = contents.getCid();
 
@@ -125,7 +124,7 @@ public class ContentsServiceImpl implements ContentsService {
 
         activeRecord.update(contents);
 
-        if (!contents.getType().equals(Types.PAGE)) {
+        if (null != contents.getType() && !contents.getType().equals(Types.PAGE)) {
             String sql = "delete from t_relationships where cid = ?";
             activeRecord.execute(sql, cid);
         }
